+++ Runtime/PostProcessing/CinemachineVolumeSettings.cs
@@ -102,14 +102,16 @@ namespace Cinemachine.PostFX
         /// <summary>
         /// Global override for the layer used by CinemachineVolumes.
         /// </summary>
+        [HideInInspector]
         [Tooltip("Global override for the layer used by CinemachineVolumes. ")]
-        public static LayerMask s_LayerMaskOverride;
+        public LayerMask m_LayerMaskOverride;
 
         /// <summary>
         /// If true, enables global layer override.
         /// </summary>
+        [HideInInspector]
         [Tooltip("If true, enables global layer override. ")]
-        public static bool s_LayerMaskOverrideEnabled;
+        public bool m_LayerMaskOverrideEnabled;
 
         class VcamExtraState
         {
@@ -255,7 +257,9 @@ namespace Cinemachine.PostFX
         {
             CameraState state = brain.CurrentCameraState;
             int numBlendables = state.NumCustomBlendables;
-            var volumes = GetDynamicBrainVolumes(brain, numBlendables);
+            var currentVolume = 
+                brain.ActiveVirtualCamera.VirtualCameraGameObject.GetComponent<CinemachineVolumeSettings>();
+            var volumes = GetDynamicBrainVolumes(brain, numBlendables, currentVolume);
             for (int i = 0; i < volumes.Count; ++i)
             {
                 volumes[i].weight = 0;
@@ -291,7 +295,8 @@ namespace Cinemachine.PostFX
 
         static string sVolumeOwnerName = "__CMVolumes";
         static  List<Volume> sVolumes = new List<Volume>();
-        static List<Volume> GetDynamicBrainVolumes(CinemachineBrain brain, int minVolumes)
+        static List<Volume> GetDynamicBrainVolumes(CinemachineBrain brain, int minVolumes, 
+            CinemachineVolumeSettings cmVolume)
         {
             // Locate the camera's child object that holds our dynamic volumes
             GameObject volumeOwner = null;
@@ -320,9 +325,9 @@ namespace Cinemachine.PostFX
                 }
 
                 // Update the volume's layer so it will be seen
-                if (s_LayerMaskOverrideEnabled)
+                if (cmVolume.m_LayerMaskOverrideEnabled)
                 {
-                    volumeOwner.layer = s_LayerMaskOverride;
+                    volumeOwner.layer = cmVolume.m_LayerMaskOverride;
                 }
                 else
                 {